Typecast sbrk pointer arithmetics into (int) when printing as KB.

+++ src/stat.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: stat.cc,v 1.367 2003/01/23 00:59:44 robertc Exp $
+ * $Id: stat.cc,v 1.368 2003/01/27 08:04:01 hno Exp $
  *
  * DEBUG: section 18    Cache Manager Statistics
  * AUTHOR: Harvest Derived
@@ -479,7 +479,7 @@ info_get(StoreEntry * sentry)
 	statCPUUsage(60));
 #if HAVE_SBRK
     storeAppendPrintf(sentry, "\tProcess Data Segment Size via sbrk(): %d KB\n",
-	((char *)sbrk(0) - (char *)sbrk_start) >> 10);
+	(int) (((char *) sbrk(0) - (char *) sbrk_start) >> 10));
 #endif
     storeAppendPrintf(sentry, "\tMaximum Resident Size: %d KB\n",
 	rusage_maxrss(&rusage));
